refactor(executor): Split simd_ops.rs into separate modules by operation type
Improves code organization and maintainability by splitting the large
simd_ops.rs (2,488 lines) into focused modules:

- simd_ops/mod.rs: PackedMask struct and shared utilities (~320 lines)
- simd_ops/aggregate.rs: Sum, min, max, count operations (~890 lines)
- simd_ops/comparison.rs: lt, gt, le, ge, eq, ne operations (~610 lines)
- simd_ops/logical.rs: Boolean mask operations (~65 lines)

No public API changes - all existing tests pass (42 tests).
