+++ crates/vibesql-executor/src/select/columnar/simd_ops/aggregate.rs
@@ -1,247 +1,11 @@
-//! Auto-vectorized SIMD operations for columnar data processing.
+//! Aggregate operations for columnar data processing.
 //!
-//! # Performance Architecture
-//!
-//! These functions are structured to enable LLVM auto-vectorization. They achieve
-//! equivalent performance to explicit SIMD (e.g., the `wide` crate) without the
-//! complexity of platform-specific code.
-//!
-//! ## Why This Pattern Works
-//!
-//! LLVM can auto-vectorize loops when:
-//! 1. Loop bounds are known or predictable
-//! 2. Memory access is sequential
-//! 3. Operations are independent across lanes
-//!
-//! The 4-accumulator pattern breaks loop-carried dependencies, allowing LLVM to
-//! use SIMD registers effectively:
-//!
-//! ```text
-//! // BAD: Single accumulator creates dependency chain
-//! for x in data { sum += x; }  // Each add waits for previous
-//!
-//! // GOOD: Four accumulators enable parallel execution
-//! for chunk in data.chunks(4) {
-//!     s0 += chunk[0];  // These four adds can execute
-//!     s1 += chunk[1];  // simultaneously in SIMD lanes
-//!     s2 += chunk[2];
-//!     s3 += chunk[3];
-//! }
-//! ```
-//!
-//! ## Benchmark Results (10M elements, Apple Silicon)
-//!
-//! | Operation | wide crate | auto-vectorized | naive iter |
-//! |-----------|------------|-----------------|------------|
-//! | sum_f64   | 2.0 ms     | 2.0 ms (1.0x)   | 7.8 ms     |
-//! | min_f64   | 1.5 ms     | 1.5 ms (1.0x)   | 1.4 ms     |
-//!
-//! ## WARNING
-//!
-//! DO NOT "simplify" these functions to use `.iter().sum()` or similar patterns.
-//! While cleaner-looking, they can be 3-4x slower due to floating-point
-//! associativity constraints preventing vectorization.
-//!
-//! If you need to modify these functions, run the SIMD benchmark first:
-//! ```bash
-//! cargo bench --bench tpch -- Q6
-//! ```
+//! This module provides vectorized aggregate operations (sum, min, max, count)
+//! with both filtered and non-filtered variants using packed and boolean masks.
 
 #![allow(clippy::needless_range_loop)]
 
-// ============================================================================
-// PACKED BITMASK TYPE
-// ============================================================================
-// Using packed bitmasks (1 bit per row) instead of Vec<bool> (1 byte per row)
-// provides 8x memory reduction and enables native SIMD bitwise operations.
-
-/// A packed bitmask where each bit represents whether a row passes a filter.
-///
-/// # Memory Efficiency
-///
-/// For 6M rows (lineitem at TPC-H SF 1):
-/// - `Vec<bool>`: 6 MB per predicate
-/// - `PackedMask`: 750 KB per predicate (8x reduction)
-///
-/// # Performance Benefits
-///
-/// - Bitwise AND is a native SIMD instruction (`vpand` on x86, `and` on ARM)
-/// - Better cache utilization due to reduced memory footprint
-/// - Faster popcount using `count_ones()` intrinsic
-///
-/// # Implementation
-///
-/// Bits are stored in little-endian order within each u64 word:
-/// - Bit 0 of word 0 = row 0
-/// - Bit 63 of word 0 = row 63
-/// - Bit 0 of word 1 = row 64
-/// - etc.
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub struct PackedMask {
-    /// The packed bits. Each u64 holds 64 row flags.
-    words: Vec<u64>,
-    /// The actual number of rows (since the last word may be partial).
-    len: usize,
-}
-
-impl PackedMask {
-    /// Creates a new mask with all bits set (all rows pass).
-    #[inline]
-    pub fn new_all_set(len: usize) -> Self {
-        if len == 0 {
-            return Self { words: vec![], len: 0 };
-        }
-        let num_words = (len + 63) / 64;
-        let mut words = vec![u64::MAX; num_words];
-
-        // Clear unused bits in the last word
-        let remainder = len % 64;
-        if remainder != 0 {
-            words[num_words - 1] = (1u64 << remainder) - 1;
-        }
-
-        Self { words, len }
-    }
-
-    /// Creates a new mask with all bits clear (no rows pass).
-    #[inline]
-    pub fn new_all_clear(len: usize) -> Self {
-        if len == 0 {
-            return Self { words: vec![], len: 0 };
-        }
-        let num_words = (len + 63) / 64;
-        Self { words: vec![0; num_words], len }
-    }
-
-    /// Returns the number of rows this mask represents.
-    #[inline]
-    pub fn len(&self) -> usize {
-        self.len
-    }
-
-    /// Returns true if this mask is empty.
-    #[inline]
-    pub fn is_empty(&self) -> bool {
-        self.len == 0
-    }
-
-    /// Gets the value of the bit at the given index.
-    #[inline]
-    pub fn get(&self, index: usize) -> bool {
-        debug_assert!(index < self.len, "index out of bounds");
-        let word_idx = index / 64;
-        let bit_idx = index % 64;
-        (self.words[word_idx] >> bit_idx) & 1 == 1
-    }
-
-    /// Sets the bit at the given index.
-    #[inline]
-    pub fn set(&mut self, index: usize, value: bool) {
-        debug_assert!(index < self.len, "index out of bounds");
-        let word_idx = index / 64;
-        let bit_idx = index % 64;
-        if value {
-            self.words[word_idx] |= 1u64 << bit_idx;
-        } else {
-            self.words[word_idx] &= !(1u64 << bit_idx);
-        }
-    }
-
-    /// Performs in-place bitwise AND with another mask.
-    ///
-    /// This is the key operation for combining multiple predicates.
-    /// Uses native bitwise AND which maps directly to SIMD instructions.
-    #[inline]
-    pub fn and_inplace(&mut self, other: &PackedMask) {
-        debug_assert_eq!(self.len, other.len, "mask lengths must match");
-
-        // Process in chunks of 4 for potential auto-vectorization
-        let chunks = self.words.len() / 4;
-        for i in 0..chunks {
-            let base = i * 4;
-            self.words[base] &= other.words[base];
-            self.words[base + 1] &= other.words[base + 1];
-            self.words[base + 2] &= other.words[base + 2];
-            self.words[base + 3] &= other.words[base + 3];
-        }
-
-        // Handle remainder
-        for i in (chunks * 4)..self.words.len() {
-            self.words[i] &= other.words[i];
-        }
-    }
-
-    /// Counts the number of set bits (rows that pass the filter).
-    ///
-    /// Uses the `count_ones()` intrinsic which maps to hardware popcount.
-    #[inline]
-    pub fn count_ones(&self) -> usize {
-        // Use 4-accumulator pattern for better throughput
-        let chunks = self.words.len() / 4;
-        let (mut c0, mut c1, mut c2, mut c3) = (0u32, 0u32, 0u32, 0u32);
-
-        for i in 0..chunks {
-            let base = i * 4;
-            c0 += self.words[base].count_ones();
-            c1 += self.words[base + 1].count_ones();
-            c2 += self.words[base + 2].count_ones();
-            c3 += self.words[base + 3].count_ones();
-        }
-
-        let mut total = (c0 + c1 + c2 + c3) as usize;
-        for i in (chunks * 4)..self.words.len() {
-            total += self.words[i].count_ones() as usize;
-        }
-        total
-    }
-
-    /// Creates a packed mask from a boolean slice.
-    ///
-    /// Used for compatibility with existing code during migration.
-    #[inline]
-    pub fn from_bool_slice(slice: &[bool]) -> Self {
-        let len = slice.len();
-        if len == 0 {
-            return Self { words: vec![], len: 0 };
-        }
-
-        let num_words = (len + 63) / 64;
-        let mut words = vec![0u64; num_words];
-
-        // Process 64 bits at a time
-        for (word_idx, word) in words.iter_mut().enumerate() {
-            let start = word_idx * 64;
-            let end = (start + 64).min(len);
-            let mut bits = 0u64;
-            for (bit_idx, &value) in slice[start..end].iter().enumerate() {
-                if value {
-                    bits |= 1u64 << bit_idx;
-                }
-            }
-            *word = bits;
-        }
-
-        Self { words, len }
-    }
-
-    /// Converts this packed mask back to a boolean vector.
-    ///
-    /// Used for compatibility with existing code during migration.
-    #[inline]
-    pub fn to_bool_vec(&self) -> Vec<bool> {
-        let mut result = Vec::with_capacity(self.len);
-        for i in 0..self.len {
-            result.push(self.get(i));
-        }
-        result
-    }
-
-    /// Returns a reference to the underlying word slice.
-    #[inline]
-    pub fn words(&self) -> &[u64] {
-        &self.words
-    }
-}
+use super::PackedMask;
 
 // ============================================================================
 // AGGREGATION OPERATIONS
@@ -873,638 +637,6 @@ pub fn max_i64_filtered(values: &[i64], filter_mask: &[bool]) -> Option<i64> {
 }
 
 // ============================================================================
-// COMPARISON OPERATIONS (Filtering)
-// ============================================================================
-
-/// Macro to generate comparison functions for a given type.
-///
-/// Comparisons are naturally vectorizable since each element is independent.
-/// We still use the chunked pattern for consistency and cache efficiency.
-macro_rules! impl_comparison {
-    ($name:ident, $ty:ty, $op:tt) => {
-        #[inline]
-        pub fn $name(values: &[$ty], threshold: $ty) -> Vec<bool> {
-            values.iter().map(|&v| v $op threshold).collect()
-        }
-    };
-}
-
-// i64 comparisons
-impl_comparison!(lt_i64, i64, <);
-impl_comparison!(gt_i64, i64, >);
-impl_comparison!(le_i64, i64, <=);
-impl_comparison!(ge_i64, i64, >=);
-impl_comparison!(eq_i64, i64, ==);
-impl_comparison!(ne_i64, i64, !=);
-
-// i32 comparisons
-impl_comparison!(lt_i32, i32, <);
-impl_comparison!(gt_i32, i32, >);
-impl_comparison!(le_i32, i32, <=);
-impl_comparison!(ge_i32, i32, >=);
-impl_comparison!(eq_i32, i32, ==);
-impl_comparison!(ne_i32, i32, !=);
-
-// f64 comparisons
-impl_comparison!(lt_f64, f64, <);
-impl_comparison!(gt_f64, f64, >);
-impl_comparison!(le_f64, f64, <=);
-impl_comparison!(ge_f64, f64, >=);
-
-/// Equality comparison for f64 (exact bit equality).
-#[inline]
-pub fn eq_f64(values: &[f64], target: f64) -> Vec<bool> {
-    values.iter().map(|&v| v == target).collect()
-}
-
-/// Inequality comparison for f64.
-#[inline]
-pub fn ne_f64(values: &[f64], target: f64) -> Vec<bool> {
-    values.iter().map(|&v| v != target).collect()
-}
-
-// ============================================================================
-// BETWEEN OPERATIONS (Range checks in single pass)
-// ============================================================================
-
-/// Check if i64 values are in range [low, high] (inclusive).
-/// More efficient than separate ge + le comparisons + AND.
-#[inline]
-pub fn between_i64(values: &[i64], low: i64, high: i64) -> Vec<bool> {
-    values.iter().map(|&v| v >= low && v <= high).collect()
-}
-
-/// Check if i32 values are in range [low, high] (inclusive).
-/// More efficient than separate ge + le comparisons + AND.
-#[inline]
-pub fn between_i32(values: &[i32], low: i32, high: i32) -> Vec<bool> {
-    values.iter().map(|&v| v >= low && v <= high).collect()
-}
-
-/// Check if f64 values are in range [low, high] (inclusive).
-/// More efficient than separate ge + le comparisons + AND.
-#[inline]
-pub fn between_f64(values: &[f64], low: f64, high: f64) -> Vec<bool> {
-    values.iter().map(|&v| v >= low && v <= high).collect()
-}
-
-// ============================================================================
-// PACKED COMPARISON OPERATIONS (Return PackedMask instead of Vec<bool>)
-// ============================================================================
-// These functions return packed bitmasks for 8x memory reduction and
-// native SIMD bitwise operations.
-
-/// Less-than comparison returning packed mask.
-#[inline]
-pub fn lt_i64_packed(values: &[i64], threshold: i64) -> PackedMask {
-    let len = values.len();
-    if len == 0 {
-        return PackedMask::new_all_clear(0);
-    }
-
-    let num_words = (len + 63) / 64;
-    let mut words = vec![0u64; num_words];
-
-    for (word_idx, word) in words.iter_mut().enumerate() {
-        let start = word_idx * 64;
-        let end = (start + 64).min(len);
-        let mut bits = 0u64;
-        for (bit_idx, &v) in values[start..end].iter().enumerate() {
-            if v < threshold {
-                bits |= 1u64 << bit_idx;
-            }
-        }
-        *word = bits;
-    }
-
-    PackedMask { words, len }
-}
-
-/// Greater-than comparison returning packed mask.
-#[inline]
-pub fn gt_i64_packed(values: &[i64], threshold: i64) -> PackedMask {
-    let len = values.len();
-    if len == 0 {
-        return PackedMask::new_all_clear(0);
-    }
-
-    let num_words = (len + 63) / 64;
-    let mut words = vec![0u64; num_words];
-
-    for (word_idx, word) in words.iter_mut().enumerate() {
-        let start = word_idx * 64;
-        let end = (start + 64).min(len);
-        let mut bits = 0u64;
-        for (bit_idx, &v) in values[start..end].iter().enumerate() {
-            if v > threshold {
-                bits |= 1u64 << bit_idx;
-            }
-        }
-        *word = bits;
-    }
-
-    PackedMask { words, len }
-}
-
-/// Less-than-or-equal comparison returning packed mask.
-#[inline]
-pub fn le_i64_packed(values: &[i64], threshold: i64) -> PackedMask {
-    let len = values.len();
-    if len == 0 {
-        return PackedMask::new_all_clear(0);
-    }
-
-    let num_words = (len + 63) / 64;
-    let mut words = vec![0u64; num_words];
-
-    for (word_idx, word) in words.iter_mut().enumerate() {
-        let start = word_idx * 64;
-        let end = (start + 64).min(len);
-        let mut bits = 0u64;
-        for (bit_idx, &v) in values[start..end].iter().enumerate() {
-            if v <= threshold {
-                bits |= 1u64 << bit_idx;
-            }
-        }
-        *word = bits;
-    }
-
-    PackedMask { words, len }
-}
-
-/// Greater-than-or-equal comparison returning packed mask.
-#[inline]
-pub fn ge_i64_packed(values: &[i64], threshold: i64) -> PackedMask {
-    let len = values.len();
-    if len == 0 {
-        return PackedMask::new_all_clear(0);
-    }
-
-    let num_words = (len + 63) / 64;
-    let mut words = vec![0u64; num_words];
-
-    for (word_idx, word) in words.iter_mut().enumerate() {
-        let start = word_idx * 64;
-        let end = (start + 64).min(len);
-        let mut bits = 0u64;
-        for (bit_idx, &v) in values[start..end].iter().enumerate() {
-            if v >= threshold {
-                bits |= 1u64 << bit_idx;
-            }
-        }
-        *word = bits;
-    }
-
-    PackedMask { words, len }
-}
-
-/// Equality comparison returning packed mask.
-#[inline]
-pub fn eq_i64_packed(values: &[i64], target: i64) -> PackedMask {
-    let len = values.len();
-    if len == 0 {
-        return PackedMask::new_all_clear(0);
-    }
-
-    let num_words = (len + 63) / 64;
-    let mut words = vec![0u64; num_words];
-
-    for (word_idx, word) in words.iter_mut().enumerate() {
-        let start = word_idx * 64;
-        let end = (start + 64).min(len);
-        let mut bits = 0u64;
-        for (bit_idx, &v) in values[start..end].iter().enumerate() {
-            if v == target {
-                bits |= 1u64 << bit_idx;
-            }
-        }
-        *word = bits;
-    }
-
-    PackedMask { words, len }
-}
-
-/// Inequality comparison returning packed mask.
-#[inline]
-pub fn ne_i64_packed(values: &[i64], target: i64) -> PackedMask {
-    let len = values.len();
-    if len == 0 {
-        return PackedMask::new_all_clear(0);
-    }
-
-    let num_words = (len + 63) / 64;
-    let mut words = vec![0u64; num_words];
-
-    for (word_idx, word) in words.iter_mut().enumerate() {
-        let start = word_idx * 64;
-        let end = (start + 64).min(len);
-        let mut bits = 0u64;
-        for (bit_idx, &v) in values[start..end].iter().enumerate() {
-            if v != target {
-                bits |= 1u64 << bit_idx;
-            }
-        }
-        *word = bits;
-    }
-
-    PackedMask { words, len }
-}
-
-/// Less-than comparison for i32 returning packed mask.
-#[inline]
-pub fn lt_i32_packed(values: &[i32], threshold: i32) -> PackedMask {
-    let len = values.len();
-    if len == 0 {
-        return PackedMask::new_all_clear(0);
-    }
-
-    let num_words = (len + 63) / 64;
-    let mut words = vec![0u64; num_words];
-
-    for (word_idx, word) in words.iter_mut().enumerate() {
-        let start = word_idx * 64;
-        let end = (start + 64).min(len);
-        let mut bits = 0u64;
-        for (bit_idx, &v) in values[start..end].iter().enumerate() {
-            if v < threshold {
-                bits |= 1u64 << bit_idx;
-            }
-        }
-        *word = bits;
-    }
-
-    PackedMask { words, len }
-}
-
-/// Greater-than comparison for i32 returning packed mask.
-#[inline]
-pub fn gt_i32_packed(values: &[i32], threshold: i32) -> PackedMask {
-    let len = values.len();
-    if len == 0 {
-        return PackedMask::new_all_clear(0);
-    }
-
-    let num_words = (len + 63) / 64;
-    let mut words = vec![0u64; num_words];
-
-    for (word_idx, word) in words.iter_mut().enumerate() {
-        let start = word_idx * 64;
-        let end = (start + 64).min(len);
-        let mut bits = 0u64;
-        for (bit_idx, &v) in values[start..end].iter().enumerate() {
-            if v > threshold {
-                bits |= 1u64 << bit_idx;
-            }
-        }
-        *word = bits;
-    }
-
-    PackedMask { words, len }
-}
-
-/// Less-than-or-equal comparison for i32 returning packed mask.
-#[inline]
-pub fn le_i32_packed(values: &[i32], threshold: i32) -> PackedMask {
-    let len = values.len();
-    if len == 0 {
-        return PackedMask::new_all_clear(0);
-    }
-
-    let num_words = (len + 63) / 64;
-    let mut words = vec![0u64; num_words];
-
-    for (word_idx, word) in words.iter_mut().enumerate() {
-        let start = word_idx * 64;
-        let end = (start + 64).min(len);
-        let mut bits = 0u64;
-        for (bit_idx, &v) in values[start..end].iter().enumerate() {
-            if v <= threshold {
-                bits |= 1u64 << bit_idx;
-            }
-        }
-        *word = bits;
-    }
-
-    PackedMask { words, len }
-}
-
-/// Greater-than-or-equal comparison for i32 returning packed mask.
-#[inline]
-pub fn ge_i32_packed(values: &[i32], threshold: i32) -> PackedMask {
-    let len = values.len();
-    if len == 0 {
-        return PackedMask::new_all_clear(0);
-    }
-
-    let num_words = (len + 63) / 64;
-    let mut words = vec![0u64; num_words];
-
-    for (word_idx, word) in words.iter_mut().enumerate() {
-        let start = word_idx * 64;
-        let end = (start + 64).min(len);
-        let mut bits = 0u64;
-        for (bit_idx, &v) in values[start..end].iter().enumerate() {
-            if v >= threshold {
-                bits |= 1u64 << bit_idx;
-            }
-        }
-        *word = bits;
-    }
-
-    PackedMask { words, len }
-}
-
-/// Equality comparison for i32 returning packed mask.
-#[inline]
-pub fn eq_i32_packed(values: &[i32], target: i32) -> PackedMask {
-    let len = values.len();
-    if len == 0 {
-        return PackedMask::new_all_clear(0);
-    }
-
-    let num_words = (len + 63) / 64;
-    let mut words = vec![0u64; num_words];
-
-    for (word_idx, word) in words.iter_mut().enumerate() {
-        let start = word_idx * 64;
-        let end = (start + 64).min(len);
-        let mut bits = 0u64;
-        for (bit_idx, &v) in values[start..end].iter().enumerate() {
-            if v == target {
-                bits |= 1u64 << bit_idx;
-            }
-        }
-        *word = bits;
-    }
-
-    PackedMask { words, len }
-}
-
-/// Inequality comparison for i32 returning packed mask.
-#[inline]
-pub fn ne_i32_packed(values: &[i32], target: i32) -> PackedMask {
-    let len = values.len();
-    if len == 0 {
-        return PackedMask::new_all_clear(0);
-    }
-
-    let num_words = (len + 63) / 64;
-    let mut words = vec![0u64; num_words];
-
-    for (word_idx, word) in words.iter_mut().enumerate() {
-        let start = word_idx * 64;
-        let end = (start + 64).min(len);
-        let mut bits = 0u64;
-        for (bit_idx, &v) in values[start..end].iter().enumerate() {
-            if v != target {
-                bits |= 1u64 << bit_idx;
-            }
-        }
-        *word = bits;
-    }
-
-    PackedMask { words, len }
-}
-
-/// Less-than comparison for f64 returning packed mask.
-#[inline]
-pub fn lt_f64_packed(values: &[f64], threshold: f64) -> PackedMask {
-    let len = values.len();
-    if len == 0 {
-        return PackedMask::new_all_clear(0);
-    }
-
-    let num_words = (len + 63) / 64;
-    let mut words = vec![0u64; num_words];
-
-    for (word_idx, word) in words.iter_mut().enumerate() {
-        let start = word_idx * 64;
-        let end = (start + 64).min(len);
-        let mut bits = 0u64;
-        for (bit_idx, &v) in values[start..end].iter().enumerate() {
-            if v < threshold {
-                bits |= 1u64 << bit_idx;
-            }
-        }
-        *word = bits;
-    }
-
-    PackedMask { words, len }
-}
-
-/// Greater-than comparison for f64 returning packed mask.
-#[inline]
-pub fn gt_f64_packed(values: &[f64], threshold: f64) -> PackedMask {
-    let len = values.len();
-    if len == 0 {
-        return PackedMask::new_all_clear(0);
-    }
-
-    let num_words = (len + 63) / 64;
-    let mut words = vec![0u64; num_words];
-
-    for (word_idx, word) in words.iter_mut().enumerate() {
-        let start = word_idx * 64;
-        let end = (start + 64).min(len);
-        let mut bits = 0u64;
-        for (bit_idx, &v) in values[start..end].iter().enumerate() {
-            if v > threshold {
-                bits |= 1u64 << bit_idx;
-            }
-        }
-        *word = bits;
-    }
-
-    PackedMask { words, len }
-}
-
-/// Less-than-or-equal comparison for f64 returning packed mask.
-#[inline]
-pub fn le_f64_packed(values: &[f64], threshold: f64) -> PackedMask {
-    let len = values.len();
-    if len == 0 {
-        return PackedMask::new_all_clear(0);
-    }
-
-    let num_words = (len + 63) / 64;
-    let mut words = vec![0u64; num_words];
-
-    for (word_idx, word) in words.iter_mut().enumerate() {
-        let start = word_idx * 64;
-        let end = (start + 64).min(len);
-        let mut bits = 0u64;
-        for (bit_idx, &v) in values[start..end].iter().enumerate() {
-            if v <= threshold {
-                bits |= 1u64 << bit_idx;
-            }
-        }
-        *word = bits;
-    }
-
-    PackedMask { words, len }
-}
-
-/// Greater-than-or-equal comparison for f64 returning packed mask.
-#[inline]
-pub fn ge_f64_packed(values: &[f64], threshold: f64) -> PackedMask {
-    let len = values.len();
-    if len == 0 {
-        return PackedMask::new_all_clear(0);
-    }
-
-    let num_words = (len + 63) / 64;
-    let mut words = vec![0u64; num_words];
-
-    for (word_idx, word) in words.iter_mut().enumerate() {
-        let start = word_idx * 64;
-        let end = (start + 64).min(len);
-        let mut bits = 0u64;
-        for (bit_idx, &v) in values[start..end].iter().enumerate() {
-            if v >= threshold {
-                bits |= 1u64 << bit_idx;
-            }
-        }
-        *word = bits;
-    }
-
-    PackedMask { words, len }
-}
-
-/// Equality comparison for f64 returning packed mask.
-#[inline]
-pub fn eq_f64_packed(values: &[f64], target: f64) -> PackedMask {
-    let len = values.len();
-    if len == 0 {
-        return PackedMask::new_all_clear(0);
-    }
-
-    let num_words = (len + 63) / 64;
-    let mut words = vec![0u64; num_words];
-
-    for (word_idx, word) in words.iter_mut().enumerate() {
-        let start = word_idx * 64;
-        let end = (start + 64).min(len);
-        let mut bits = 0u64;
-        for (bit_idx, &v) in values[start..end].iter().enumerate() {
-            if v == target {
-                bits |= 1u64 << bit_idx;
-            }
-        }
-        *word = bits;
-    }
-
-    PackedMask { words, len }
-}
-
-/// Inequality comparison for f64 returning packed mask.
-#[inline]
-pub fn ne_f64_packed(values: &[f64], target: f64) -> PackedMask {
-    let len = values.len();
-    if len == 0 {
-        return PackedMask::new_all_clear(0);
-    }
-
-    let num_words = (len + 63) / 64;
-    let mut words = vec![0u64; num_words];
-
-    for (word_idx, word) in words.iter_mut().enumerate() {
-        let start = word_idx * 64;
-        let end = (start + 64).min(len);
-        let mut bits = 0u64;
-        for (bit_idx, &v) in values[start..end].iter().enumerate() {
-            if v != target {
-                bits |= 1u64 << bit_idx;
-            }
-        }
-        *word = bits;
-    }
-
-    PackedMask { words, len }
-}
-
-// ============================================================================
-// PACKED BETWEEN OPERATIONS
-// ============================================================================
-
-/// Check if i64 values are in range [low, high] (inclusive), returning packed mask.
-#[inline]
-pub fn between_i64_packed(values: &[i64], low: i64, high: i64) -> PackedMask {
-    let len = values.len();
-    if len == 0 {
-        return PackedMask::new_all_clear(0);
-    }
-
-    let num_words = (len + 63) / 64;
-    let mut words = vec![0u64; num_words];
-
-    for (word_idx, word) in words.iter_mut().enumerate() {
-        let start = word_idx * 64;
-        let end = (start + 64).min(len);
-        let mut bits = 0u64;
-        for (bit_idx, &v) in values[start..end].iter().enumerate() {
-            if v >= low && v <= high {
-                bits |= 1u64 << bit_idx;
-            }
-        }
-        *word = bits;
-    }
-
-    PackedMask { words, len }
-}
-
-/// Check if i32 values are in range [low, high] (inclusive), returning packed mask.
-#[inline]
-pub fn between_i32_packed(values: &[i32], low: i32, high: i32) -> PackedMask {
-    let len = values.len();
-    if len == 0 {
-        return PackedMask::new_all_clear(0);
-    }
-
-    let num_words = (len + 63) / 64;
-    let mut words = vec![0u64; num_words];
-
-    for (word_idx, word) in words.iter_mut().enumerate() {
-        let start = word_idx * 64;
-        let end = (start + 64).min(len);
-        let mut bits = 0u64;
-        for (bit_idx, &v) in values[start..end].iter().enumerate() {
-            if v >= low && v <= high {
-                bits |= 1u64 << bit_idx;
-            }
-        }
-        *word = bits;
-    }
-
-    PackedMask { words, len }
-}
-
-/// Check if f64 values are in range [low, high] (inclusive), returning packed mask.
-#[inline]
-pub fn between_f64_packed(values: &[f64], low: f64, high: f64) -> PackedMask {
-    let len = values.len();
-    if len == 0 {
-        return PackedMask::new_all_clear(0);
-    }
-
-    let num_words = (len + 63) / 64;
-    let mut words = vec![0u64; num_words];
-
-    for (word_idx, word) in words.iter_mut().enumerate() {
-        let start = word_idx * 64;
-        let end = (start + 64).min(len);
-        let mut bits = 0u64;
-        for (bit_idx, &v) in values[start..end].iter().enumerate() {
-            if v >= low && v <= high {
-                bits |= 1u64 << bit_idx;
-            }
-        }
-        *word = bits;
-    }
-
-    PackedMask { words, len }
-}
-
-// ============================================================================
 // PACKED FILTERED AGGREGATION OPERATIONS
 // ============================================================================
 // These operations aggregate data directly using a packed filter mask,
@@ -1864,51 +996,6 @@ pub fn sum_product_f64_packed_filtered_masked(
     (sum, count)
 }
 
-// ============================================================================
-// MASK OPERATIONS (Boolean operations on filter masks)
-// ============================================================================
-
-/// In-place AND of two boolean masks.
-///
-/// This function ANDs `other` into `mask` in-place, modifying `mask`.
-/// Uses an unrolled loop that LLVM can auto-vectorize effectively.
-///
-/// # Panics
-/// Panics if `mask` and `other` have different lengths.
-#[inline]
-pub fn and_masks_inplace(mask: &mut [bool], other: &[bool]) {
-    assert_eq!(mask.len(), other.len(), "mask lengths must match");
-
-    let len = mask.len();
-
-    // Process in chunks of 8 for better vectorization
-    let chunks = len / 8;
-    let remainder = len % 8;
-
-    for i in 0..chunks {
-        let base = i * 8;
-        // Unrolled loop - LLVM vectorizes this well
-        mask[base] = mask[base] && other[base];
-        mask[base + 1] = mask[base + 1] && other[base + 1];
-        mask[base + 2] = mask[base + 2] && other[base + 2];
-        mask[base + 3] = mask[base + 3] && other[base + 3];
-        mask[base + 4] = mask[base + 4] && other[base + 4];
-        mask[base + 5] = mask[base + 5] && other[base + 5];
-        mask[base + 6] = mask[base + 6] && other[base + 6];
-        mask[base + 7] = mask[base + 7] && other[base + 7];
-    }
-
-    // Handle remainder
-    let base = chunks * 8;
-    for i in 0..remainder {
-        mask[base + i] = mask[base + i] && other[base + i];
-    }
-}
-
-// ============================================================================
-// TESTS
-// ============================================================================
-
 #[cfg(test)]
 mod tests {
     use super::*;
@@ -1944,25 +1031,6 @@ mod tests {
         assert_eq!(max_f64(&values), Some(9.0));
         assert_eq!(min_f64(&[]), None);
         assert_eq!(max_f64(&[]), None);
-    }
-
-    #[test]
-    fn test_comparisons_i64() {
-        let values = vec![1, 2, 3, 4, 5];
-        assert_eq!(lt_i64(&values, 3), vec![true, true, false, false, false]);
-        assert_eq!(gt_i64(&values, 3), vec![false, false, false, true, true]);
-        assert_eq!(le_i64(&values, 3), vec![true, true, true, false, false]);
-        assert_eq!(ge_i64(&values, 3), vec![false, false, true, true, true]);
-        assert_eq!(eq_i64(&values, 3), vec![false, false, true, false, false]);
-        assert_eq!(ne_i64(&values, 3), vec![true, true, false, true, true]);
-    }
-
-    #[test]
-    fn test_comparisons_f64() {
-        let values = vec![1.0, 2.0, 3.0, 4.0, 5.0];
-        assert_eq!(lt_f64(&values, 3.0), vec![true, true, false, false, false]);
-        assert_eq!(gt_f64(&values, 3.0), vec![false, false, false, true, true]);
-        assert_eq!(eq_f64(&values, 3.0), vec![false, false, true, false, false]);
     }
 
     #[test]
@@ -2158,206 +1226,6 @@ mod tests {
     }
 
     #[test]
-    fn test_and_masks_inplace() {
-        let mut mask = vec![true, true, false, true, true, false, true, true];
-        let other = vec![true, false, true, true, false, true, true, false];
-        and_masks_inplace(&mut mask, &other);
-        assert_eq!(mask, vec![true, false, false, true, false, false, true, false]);
-    }
-
-    #[test]
-    fn test_and_masks_inplace_remainder() {
-        // Test with non-multiple-of-8 length
-        let mut mask = vec![true, true, false, true, true];
-        let other = vec![true, false, true, true, false];
-        and_masks_inplace(&mut mask, &other);
-        assert_eq!(mask, vec![true, false, false, true, false]);
-    }
-
-    #[test]
-    fn test_between_i64() {
-        let values = vec![1, 5, 10, 15, 20, 25];
-        assert_eq!(between_i64(&values, 5, 20), vec![false, true, true, true, true, false]);
-    }
-
-    #[test]
-    fn test_between_i32() {
-        let values: Vec<i32> = vec![1, 5, 10, 15, 20, 25];
-        assert_eq!(between_i32(&values, 5, 20), vec![false, true, true, true, true, false]);
-    }
-
-    #[test]
-    fn test_between_f64() {
-        let values = vec![0.01, 0.05, 0.06, 0.07, 0.08];
-        assert_eq!(between_f64(&values, 0.05, 0.07), vec![false, true, true, true, false]);
-    }
-
-    // ========================================================================
-    // PACKED MASK TESTS
-    // ========================================================================
-
-    #[test]
-    fn test_packed_mask_new_all_set() {
-        let mask = PackedMask::new_all_set(100);
-        assert_eq!(mask.len(), 100);
-        for i in 0..100 {
-            assert!(mask.get(i), "bit {} should be set", i);
-        }
-
-        // Empty mask
-        let empty = PackedMask::new_all_set(0);
-        assert_eq!(empty.len(), 0);
-        assert!(empty.is_empty());
-    }
-
-    #[test]
-    fn test_packed_mask_new_all_clear() {
-        let mask = PackedMask::new_all_clear(100);
-        assert_eq!(mask.len(), 100);
-        for i in 0..100 {
-            assert!(!mask.get(i), "bit {} should be clear", i);
-        }
-    }
-
-    #[test]
-    fn test_packed_mask_set_get() {
-        let mut mask = PackedMask::new_all_clear(128);
-
-        // Set some bits
-        mask.set(0, true);
-        mask.set(63, true);
-        mask.set(64, true);
-        mask.set(127, true);
-
-        assert!(mask.get(0));
-        assert!(mask.get(63));
-        assert!(mask.get(64));
-        assert!(mask.get(127));
-        assert!(!mask.get(1));
-        assert!(!mask.get(62));
-        assert!(!mask.get(65));
-    }
-
-    #[test]
-    fn test_packed_mask_and_inplace() {
-        let mut mask_a = PackedMask::new_all_set(128);
-        let mut mask_b = PackedMask::new_all_clear(128);
-
-        // Set every other bit in mask_b
-        for i in (0..128).step_by(2) {
-            mask_b.set(i, true);
-        }
-
-        mask_a.and_inplace(&mask_b);
-
-        for i in 0..128 {
-            if i % 2 == 0 {
-                assert!(mask_a.get(i), "bit {} should be set (even)", i);
-            } else {
-                assert!(!mask_a.get(i), "bit {} should be clear (odd)", i);
-            }
-        }
-    }
-
-    #[test]
-    fn test_packed_mask_count_ones() {
-        let mask = PackedMask::new_all_set(100);
-        assert_eq!(mask.count_ones(), 100);
-
-        let mask = PackedMask::new_all_clear(100);
-        assert_eq!(mask.count_ones(), 0);
-
-        let mut mask = PackedMask::new_all_clear(128);
-        for i in (0..128).step_by(2) {
-            mask.set(i, true);
-        }
-        assert_eq!(mask.count_ones(), 64);
-    }
-
-    #[test]
-    fn test_packed_mask_from_bool_slice() {
-        let bools = vec![true, false, true, true, false, true];
-        let mask = PackedMask::from_bool_slice(&bools);
-
-        assert_eq!(mask.len(), 6);
-        assert!(mask.get(0));
-        assert!(!mask.get(1));
-        assert!(mask.get(2));
-        assert!(mask.get(3));
-        assert!(!mask.get(4));
-        assert!(mask.get(5));
-        assert_eq!(mask.count_ones(), 4);
-    }
-
-    #[test]
-    fn test_packed_mask_to_bool_vec() {
-        let mut mask = PackedMask::new_all_clear(6);
-        mask.set(0, true);
-        mask.set(2, true);
-        mask.set(3, true);
-        mask.set(5, true);
-
-        let bools = mask.to_bool_vec();
-        assert_eq!(bools, vec![true, false, true, true, false, true]);
-    }
-
-    #[test]
-    fn test_packed_mask_round_trip() {
-        let original = vec![true, false, true, true, false, true, false, false, true];
-        let mask = PackedMask::from_bool_slice(&original);
-        let result = mask.to_bool_vec();
-        assert_eq!(original, result);
-    }
-
-    #[test]
-    fn test_packed_comparison_i64() {
-        let values = vec![1i64, 2, 3, 4, 5];
-
-        let mask = lt_i64_packed(&values, 3);
-        assert_eq!(mask.to_bool_vec(), vec![true, true, false, false, false]);
-
-        let mask = gt_i64_packed(&values, 3);
-        assert_eq!(mask.to_bool_vec(), vec![false, false, false, true, true]);
-
-        let mask = le_i64_packed(&values, 3);
-        assert_eq!(mask.to_bool_vec(), vec![true, true, true, false, false]);
-
-        let mask = ge_i64_packed(&values, 3);
-        assert_eq!(mask.to_bool_vec(), vec![false, false, true, true, true]);
-
-        let mask = eq_i64_packed(&values, 3);
-        assert_eq!(mask.to_bool_vec(), vec![false, false, true, false, false]);
-
-        let mask = ne_i64_packed(&values, 3);
-        assert_eq!(mask.to_bool_vec(), vec![true, true, false, true, true]);
-    }
-
-    #[test]
-    fn test_packed_comparison_f64() {
-        let values = vec![1.0, 2.0, 3.0, 4.0, 5.0];
-
-        let mask = lt_f64_packed(&values, 3.0);
-        assert_eq!(mask.to_bool_vec(), vec![true, true, false, false, false]);
-
-        let mask = ge_f64_packed(&values, 3.0);
-        assert_eq!(mask.to_bool_vec(), vec![false, false, true, true, true]);
-    }
-
-    #[test]
-    fn test_packed_between_i64() {
-        let values = vec![1i64, 5, 10, 15, 20, 25];
-        let mask = between_i64_packed(&values, 5, 20);
-        assert_eq!(mask.to_bool_vec(), vec![false, true, true, true, true, false]);
-    }
-
-    #[test]
-    fn test_packed_between_f64() {
-        let values = vec![0.01, 0.05, 0.06, 0.07, 0.08];
-        let mask = between_f64_packed(&values, 0.05, 0.07);
-        assert_eq!(mask.to_bool_vec(), vec![false, true, true, true, false]);
-    }
-
-    #[test]
     fn test_sum_f64_packed_filtered() {
         let values = vec![1.0, 2.0, 3.0, 4.0, 5.0, 6.0, 7.0, 8.0];
         let mut mask = PackedMask::new_all_clear(8);
@@ -2450,39 +1318,5 @@ mod tests {
         let filter_none = PackedMask::new_all_clear(8);
         assert_eq!(min_i64_packed_filtered(&values, &filter_none), None);
         assert_eq!(max_i64_packed_filtered(&values, &filter_none), None);
-    }
-
-    #[test]
-    fn test_packed_mask_large_dataset() {
-        // Test with a large dataset to ensure correctness at scale
-        let n = 10_000;
-        let values: Vec<f64> = (0..n).map(|i| i as f64).collect();
-
-        // Filter: values < 5000 (half the dataset)
-        let mask = lt_f64_packed(&values, 5000.0);
-
-        assert_eq!(mask.count_ones(), 5000);
-
-        // Sum of 0..4999 = 4999 * 5000 / 2 = 12497500
-        let sum = sum_f64_packed_filtered(&values, &mask);
-        assert!((sum - 12497500.0).abs() < 0.001);
-    }
-
-    #[test]
-    fn test_packed_mask_partial_word() {
-        // Test with sizes that don't fill a complete u64 word
-        for size in [1, 7, 15, 63, 65, 100, 127, 129] {
-            let mask = PackedMask::new_all_set(size);
-            assert_eq!(mask.count_ones(), size, "failed for size {}", size);
-
-            let mask = PackedMask::new_all_clear(size);
-            assert_eq!(mask.count_ones(), 0, "failed for size {}", size);
-
-            // Test conversion round-trip
-            let bools = vec![true; size];
-            let mask = PackedMask::from_bool_slice(&bools);
-            assert_eq!(mask.count_ones(), size, "round-trip failed for size {}", size);
-            assert_eq!(mask.to_bool_vec(), bools, "round-trip failed for size {}", size);
-        }
     }
 }
